+++ Xocolatl/CocoaHTTPServer/Core/Mime/MultipartMessageHeaderField.m
@@ -6,15 +6,15 @@
 #pragma mark log level
 
 #ifdef DEBUG
-static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
+static const NSInteger httpLogLevel = HTTP_LOG_LEVEL_WARN;
 #else
-static const int httpLogLevel = HTTP_LOG_LEVEL_WARN;
+static const NSInteger httpLogLevel = HTTP_LOG_LEVEL_WARN;
 #endif
 
 
 // helpers
-int findChar(const char* str,int length, char c);
-NSString* extractParamValue(const char* bytes, int length, NSStringEncoding encoding);
+NSInteger findChar(const char* str,NSInteger length, char c);
+NSString* extractParamValue(const char* bytes, NSInteger length, NSStringEncoding encoding);
 
 //-----------------------------------------------------------------
 // interface MultipartMessageHeaderField (private)
@@ -22,7 +22,7 @@ NSString* extractParamValue(const char* bytes, int length, NSStringEncoding enco
 
 
 @interface MultipartMessageHeaderField (private)
--(BOOL) parseHeaderValueBytes:(char*) bytes length:(int) length encoding:(NSStringEncoding) encoding;
+-(BOOL) parseHeaderValueBytes:(char*) bytes length:(NSInteger) length encoding:(NSStringEncoding) encoding;
 @end
 
 
@@ -38,9 +38,9 @@ NSString* extractParamValue(const char* bytes, int length, NSStringEncoding enco
 	params = [[NSMutableDictionary alloc] initWithCapacity:1];
 
 	char* bytes = (char*)data.bytes;
-	int length = data.length;
+	NSInteger length = data.length;
 
-	int separatorOffset = findChar(bytes, length, ':');
+	NSInteger separatorOffset = findChar(bytes, length, ':');
 	if( (-1 == separatorOffset) || (separatorOffset >= length-2) ) {
 		HTTPLogError(@"MultipartFormDataParser: Bad format.No colon in field header.");
 		// tear down
@@ -90,8 +90,8 @@ NSString* extractParamValue(const char* bytes, int length, NSStringEncoding enco
 }
 
 
--(BOOL) parseHeaderValueBytes:(char*) bytes length:(int) length encoding:(NSStringEncoding) encoding {
-	int offset = 0;
+-(BOOL) parseHeaderValueBytes:(char*) bytes length:(NSInteger) length encoding:(NSStringEncoding) encoding {
+	NSInteger offset = 0;
 	NSString* currentParam = nil;
 	BOOL insideQuote = NO;
 	while( offset < length ) {
@@ -182,8 +182,8 @@ NSString* extractParamValue(const char* bytes, int length, NSStringEncoding enco
 @end
 
 
-int findChar(const char* str,int length, char c) {
-	int offset = 0;
+NSInteger findChar(const char* str,NSInteger length, char c) {
+	NSInteger offset = 0;
 	while( offset < length ) {
 		if( str[offset] == c )
 			return offset;
@@ -193,7 +193,7 @@ int findChar(const char* str,int length, char c) {
 }
 
 
-NSString* extractParamValue(const char* bytes, int length, NSStringEncoding encoding) {
+NSString* extractParamValue(const char* bytes, NSInteger length, NSStringEncoding encoding) {
 	if( !length ) 
 		return nil;
 	NSMutableString* value = nil;
